requirements.txt
requests==2.31.0
openai
azure-search-documents==11.4.0a20230509004 
azure-storage-blob==12.14.1
python-dotenv==1.0.0
semantic-kernel
asyncio
streamlit